requirements.txt
pandas>=2.0.0
pyyaml>=6.0
requests>=2.31.0
sentence-transformers>=2.2.0
numpy>=1.21.0
